Restore types
Partially reverts https://github.com/TrinityCore/TrinityCore/pull/19277/commits/cd2534771ea151b10f43997d68799487eb5b3b5e

+++ src/server/game/DataStores/DB2Structure.h
@@ -2071,8 +2071,8 @@ struct RewardPackEntry
     uint32 ID;
     uint32 Money;
     float ArtifactXPMultiplier;
-    uint32 ArtifactXPDifficulty;
-    uint32 ArtifactCategoryID;
+    uint8 ArtifactXPDifficulty;
+    uint8 ArtifactCategoryID;
     uint32 TitleID;
     uint32 Unused;
 };
